Solution to Count Trailing Zeros in a Binary Number

+++ Count Trailing Zeros in a Binary Number@@67882a731b4392d8fc1577f1/code.c
@@ -7,6 +7,6 @@
 int main() {
     int a ;
     scanf("%d" , &a);
-    printf("%d", __builtin_clz(a));
+    printf("%d", __builtin_clz((unsigned int)a));
     return 0;
 }
